# burn-link-backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/burn-link:
    post:
      description: Create new burn link
      operationId: BurnLinkController_createBurnLink
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBurnLinkRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBurnLinkResponse'
      summary: ''
  /api/burn-link/{id}:
    get:
      description: Get burn-link content
      operationId: BurnLinkController_getBurnLink
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBurnLinkResponse'
      summary: ''
    delete:
      description: Delete burn-link
      operationId: BurnLinkController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      summary: ''
...
info:
  title: HUB API
  description: ''
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: jwt
      description: JWT cookie
  schemas:
...
    CreateBurnLinkRequest:
      type: object
      properties:
        message:
          type: string
        expires:
          format: date-time
          type: string
      required:
        - message
    CreateBurnLinkResponse:
      type: object
      properties:
        id:
          type: number
      required:
        - id
    GetBurnLinkResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...